Execution Context
In this topic, I delve into the concept of execution contexts in JavaScript. I'll study the Global Execution Context, Function Execution Context, as well as the phases of creation and execution. Understanding these contexts and phases is fundamental to comprehending how JavaScript code is processed, and it forms the basis for advanced programming techniques and debugging.